requirements.txt
PyInquirer==1.0.3
PyYAML==5.1
Pygments==2.3.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
pip==19.0.3
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
regex==2019.4.14
requests==2.21.0
setuptools==40.8.0
six==1.12.0
urllib3==1.24.1
wcwidth==0.1.7